# server/project-management/src/main/kotlin/com/example/projectmanagement/models/Project.kt
package com.example.projectmanagement.models

import jakarta.persistence.Entity
import jakarta.persistence.GeneratedValue
import jakarta.persistence.GenerationType
import jakarta.persistence.Id
import jakarta.persistence.Table

@Entity
@Table(name = "projects")
data class Project(
    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
    val id: Long = 0,
    val author: Long,
    val date: String,
    val name: String
)